Refactor booking methods using factory and cleaning tests

+++ tests/restful-booker/booking/delete-booking.test.ts
@@ -1,23 +1,36 @@
 import { AuthClient } from '../../../src/restful-booker/clients/AuthClient';
 import { BookingClient } from '../../../src/restful-booker/clients/BookingClient';
+import { BookingFactory } from '../../../src/factories/booking.factory';
 
 describe('DELETE /booking/:id', () => {
   const bookingClient = new BookingClient();
+  let bookingId: number | undefined;
+  const authClient = new AuthClient();
+  let token: string;
+  let bookingDeleted = false;
+
+  beforeEach(async () => {
+    const authResponse = await authClient.createToken({
+      username: 'admin',
+      password: 'password123'
+    });
+
+    token = authResponse.data.token;
+    const booking = BookingFactory.createBooking();
+    const response = await bookingClient.createBooking(booking);
+    bookingId = response.data.bookingid;
+  });
+
+  afterEach(async () => {
+    if (bookingId && !bookingDeleted) {
+      await bookingClient.deleteBooking(bookingId, token);
+    }
+  });
 
   test('should delete an existing booking', async () => {
-    const authClient = new AuthClient();
+    const response = await bookingClient.deleteBooking(bookingId!, token);
 
-    const authResponse = await authClient.createToken({
-            username: 'admin',
-            password: 'password123'
-        });
-
-        const token = authResponse.data.token;
-
-        const bookingId = 1;
-
-        const response = await bookingClient.deleteBooking(bookingId, token);
-
-        expect(response.status).toBe(201);
+    expect(response.status).toBe(201);
+    bookingDeleted = true;
   });
 });
